validators in controllers(applicationcontroller is under construction)

+++ app/Http/Controllers/SelectOptionController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\SelectOption;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Facades\Response;
 
 class SelectOptionController extends Controller
 {
@@ -14,20 +16,43 @@ class SelectOptionController extends Controller
    }
 
     public function store(Request $request) {
-
+        $validator = Validator::make($request->all(),[
+            'text' => ['required', 'min:3'],
+        ]);
+        if ($validator->fails()) {
+          return Response::json(array(
+              'error' => $validator->getMessageBag()->toArray()
+          ), 400);
+        }
         $item = new SelectOption;
         $item->name = $request->text;
         $item->save();
-    }  // HERE!!!
+    }
 
     public function delete(Request $request) {
+        $validator = Validator::make($request->all(),[
+            'id' => ['required'],
+        ]);
+        if ($validator->fails()) {
+          return Response::json(array(
+              'error' => $validator->getMessageBag()->toArray()
+          ), 400);
+        }
         SelectOption::where('id', $request->id)->delete();
-    }  // HERE!!!
+    }
 
     public function update(Request $request) {
-
+        $validator = Validator::make($request->all(),[
+            'id' => ['required'],
+            'value' => ['required', 'min:3'],
+        ]);
+        if ($validator->fails()) {
+          return Response::json(array(
+              'error' => $validator->getMessageBag()->toArray()
+          ), 400);
+        }
         $item = SelectOption::find($request->id);
         $item->name = $request->value;
         $item->update();
-    }  // HERE!!!
+    }
 }
